Optimize SQL for account delete
Type: generic-bugfix

Why is this change necessary?
We were getting intermittent integrity errors and timeouts when
deleting accounts with many items and issues. This appears to be
caused by SQL Alchemy's method of handling cascading deletes,
which is inefficient and does not appear to handle transactional
locks well, allowing for race conditions.

This change addresses the need by:
Deleting accounts and related records with a raw sql query