Add bulk enable and disable account service
Type: feature

Why is this change necessary?
As the number of accounts being watched increases, it becomes harder to
manage different environments and which accounts they should be
watching. This service enables an API and command-line tool (through
manage.py) that allows a bulk list of accounts to be enabled or disabled
in one command.