Add Bandit gate to Security Monkey (#688) :sleuth_or_spy:
* Addressing Bandit findings

This commit addresses a few Bandit findings so that we can run
a gate in the future.  A few issues are marked non dangerous
(use of md5 for non-security), and a couple issues were fixed
(autoescaping Jinja and using yaml.safe_load).

* Update travis to run Bandit

This commit adds a Bandit gate to travis to run against each
commit.  Bandit will be run with medium+ severity and confidence
filters and exclude the tests folder.